// app/writing/panel-with-custom-order/_componnets/fields/interfaces.ts
export enum FIELD {
  POSITION = 'element:settings:position',
  LAYOUT = 'element:settings:layout',
  APPEARANCE = 'element:settings:appearance',
  FILL = 'element:settings:fill',
  STROKE = 'element:settings:stroke',
  EFFECTS = 'elements:settings:effects',
  EXPORT = 'elements:settings:export',
}

export enum LAYOUT_ALIGN {
  LEFT = 'layout:align:left',
  RIGHT = 'layout:align:right',
  TOP = 'layout:align:top',
  BOTTOM = 'layout:align:bottom',
  CENTER = 'layout:align:center',
  MIDDLE = 'layout:align:middle',
}

export enum LAYOUT_ALIGN_CATEGORY {
  HORIZONTAL = 'layout:align:category:horizontal',
  VERTICAL = 'layout:align:category:vertical',
}